requirements.txt
numpy
scikit-learn
torch
torchvision
Pillow
